Fix build problems on OSX
The CMake toolchain already defines a value for `${CMAKE_C_FLAGS}`, so
just calling `set()` on it overwrites these flags.  This is critical on
any platform that uses `clang` (such as `OSX`) because those platforms
require passing the `-target` argument to `clang`, as otherwise `clang`
is configured to output `x86_64-linux-gnu` object code.  This patch
causes the build script to append to this variable, instead of replacing
it.

This also follows the more typical pattern of putting all
CMake-generated output within a `build` subdirectory.  This is not
essential, but was my first step in separating out what was and was not
important when building.